fix: keep the reconciliation port inside the repo's lint ceilings
The new scan/seal/verify code pushed three files past the 2500-line limit
and drifted two generated artifacts. None of this changes behaviour.

canonical_index.c (2055 -> 2522) is split: the read surface (project
listing, symbol lookup, blast radius, structure, caller search, stats,
code search) moves to canonical_index_query.c, leaving the write/scan
surface behind. The two share ci_conn / ci_resolve_project_id /
ci_resolve_file_id plus CI_ERRBUF / CI_MAX_DEPS through a new private
canonical_index_internal.h. The query surface was the cheaper cut: it
needed three file-local helpers hoisted, against eight for the scan block.

server.c was already exactly at 2500, so any new route broke it. Its
dispatch rows move to server/server_dispatch_defs_data.h, the same split
server/cli_dispatch_defs_data.h already uses. check-dispatch-caps.py and
check-v1-method-coverage.py both parsed the table out of server.c and now
read the rows wherever they live; verified against their own plant tests.

handle_index_verify moves from server_state.c to server_state_index.c,
where that file's own comment already says the index handlers live.

Regenerates tests/baselines/db2/declarations-v1.json and applies
clang-format.
