+++ CMakeLists.txt
@@ -436,6 +436,7 @@ set(DB2_SRCS
     ${AIMEE_SRC_DIR}/modules/db2/c/agent_hints.c
     ${AIMEE_SRC_DIR}/modules/db2/c/agent_outcomes.c
     ${AIMEE_SRC_DIR}/modules/db2/c/canonical_index.c
+    ${AIMEE_SRC_DIR}/modules/db2/c/canonical_index_query.c
     ${AIMEE_SRC_DIR}/modules/db2/c/anti_patterns.c
     ${AIMEE_SRC_DIR}/modules/db2/c/collab_rules.c
     ${AIMEE_SRC_DIR}/modules/db2/c/curiosity.c
