+++ src/server/server.c
@@ -1388,268 +1388,7 @@ static int handle_memory_user_capture(server_ctx_t *ctx, server_conn_t *conn, cJ
 }
 
 static const server_method_dispatch_t server_dispatch_table[] = {
-    /* Server */
-    {"server.info", handle_server_info},
-    {"server.health", handle_server_health},
-    {"api.status", handle_api_status},
-    {"api.enable", handle_api_enable},
-    {"api.rotate_bearer", handle_api_rotate_bearer},
-    {"api.enroll_bearer", handle_api_enroll_bearer},
-    {"api.disable", handle_api_disable},
-    {"insights.overview", handle_insights_overview},
-    {"toolset.list", handle_toolset_list},
-    {"toolset.show", handle_toolset_show},
-    {"toolset.resolve", handle_toolset_resolve},
-    {"worktree.gc", handle_worktree_gc},
-    {"delegate.backend_list", handle_delegate_backend_list},
-    {"delegate.backend_exec", handle_delegate_backend_exec},
-    {"delegate.sandbox_list", handle_delegate_sandbox_list},
-    {"delegate.sandbox_gc", handle_delegate_sandbox_gc},
-    {"provider.list", handle_provider_list},
-    {"provider.show", handle_provider_show},
-    {"provider.models", handle_provider_models},
-    {"provider.test", handle_provider_test},
-    {"provider.quota", handle_provider_quota},
-    {"provider.get", handle_provider_get},
-    {"provider.set", handle_provider_set},
-    {"catalog.list", handle_model_list},
-    {"catalog.show", handle_model_show},
-    {"catalog.refresh", handle_model_refresh},
-    {"init.run", handle_init_run},
-    {"launch.run", handle_launch_run},
-    {"hud.status", handle_hud_status},
-    {"model.list", handle_agent_list},
-    {"model.add", handle_agent_add},
-    {"model.local", handle_agent_local},
-    {"model.remove", handle_agent_remove},
-    {"model.enable", handle_agent_enable},
-    {"model.roles", handle_agent_roles},
-    {"model.personas", handle_agent_personas},
-    {"model.set", handle_agent_set},
-    {"model.disable", handle_agent_disable},
-    {"model.probe", handle_agent_probe},
-    {"model.stats", handle_agent_stats},
-    {"model.draft", handle_agent_draft},
-    {"model.setup", handle_agent_setup},
-    {"model.setup_poll", handle_agent_setup_poll},
-    {"model.cli_oauth_start", handle_agent_cli_oauth_start},
-    {"model.cli_oauth_code", handle_agent_cli_oauth_code},
-    {"model.cli_oauth_poll", handle_agent_cli_oauth_poll},
-    {"hooks.pre", handle_hooks_pre},
-    {"hooks.post", handle_hooks_post},
-    {"hooks.session_start", handle_hooks_session_start},
-    {"session.brief_assemble", handle_session_brief_assemble},
-    {"memory.user_capture", handle_memory_user_capture},
-    {"session.create", handle_session_create},
-    {"session.record_transcript", handle_session_record_transcript},
-    {"session.list", handle_session_list},
-    {"session.get", handle_session_get},
-    {"session.close", handle_session_close},
-    {"chat.graceful_cancel", handle_chat_graceful_cancel},
-    {"chat.interrupt", handle_chat_interrupt},
-    {"session.brief", handle_session_brief},
-    {"session.attach", handle_session_attach},
-    {"session.detach", handle_session_detach},
-    {"session.presence", handle_session_presence},
-    {"trajectory.export", handle_trajectory_export},
-    {"trajectory.batch", handle_trajectory_batch},
-    /* Memory */
-    {"memory.search", handle_memory_search},
-    {"memory.store", handle_memory_store},
-    {"memory.list", handle_memory_list},
-    {"memory.review_list", handle_memory_review_list},
-    {"memory.reject", handle_memory_reject},
-    {"memory.restore", handle_memory_restore},
-    {"memory.stats", handle_memory_stats},
-    {"memory.get", handle_memory_get},
-    {"memory.delete", handle_memory_delete},
-    {"memory.supersede", handle_memory_supersede},
-    {"facts.retract", handle_facts_retract},
-    {"entities.merge", handle_entities_merge},
-    {"entities.unmerge", handle_entities_unmerge},
-    {"memory.read", handle_memory_read},
-    {"memory.benchmark", handle_memory_benchmark},
-    {"index.scan", handle_index_scan},
-    {"index.verify", handle_index_verify},
-    {"index.ingest", handle_index_ingest},
-    {"index.find", handle_index_find},
-    {"index.list", handle_index_list},
-    {"index.blast_radius", handle_index_blast_radius},
-    {"index.structure", handle_index_structure},
-    {"index.span", handle_index_span},
-    {"index.investigate", handle_index_investigate},
-    {"index.hybrid", handle_index_hybrid},
-    {"index.find_callers", handle_index_find_callers},
-    {"index.deps", handle_index_deps},
-    {"graph.sync_code", handle_graph_sync_code},
-    {"graph.explain", handle_graph_explain},
-    {"blast_radius.preview", handle_blast_radius_preview},
-    /* KB */
-    {"kb.search", handle_kb_search},
-    {"evidence.trace_retrieval_event", handle_evidence_trace},
-    {"evidence.provenance_retrieval_event", handle_evidence_provenance},
-    {"evidence.fidelity_retrieval_event", handle_evidence_fidelity},
-    {"css.signals", handle_css_signals},
-    {"curator.implements", handle_curator_implements},
-    {"curator.synthesize", handle_curator_synthesize},
-    {"curator.contradictions", handle_curator_contradictions},
-    {"curator.invalidated", handle_curator_invalidated},
-    {"kb.build", handle_kb_build},
-    {"kb.update", handle_kb_update},
-    {"kb.docs.push", handle_kb_docs_push},
-    {"kb.ingest", handle_kb_ingest},
-    {"kb.ingest.status", handle_kb_ingest_status},
-    {"kb.reembed", handle_kb_reembed},
-    {"memory.embed", handle_memory_embed},
-    {"kb.status", handle_kb_status},
-    {"optimize.export", handle_optimize_export},
-    {"optimize.promote", handle_optimize_promote},
-    {"optimize.replay_record", handle_optimize_replay_record},
-    {"calibration.readiness", handle_calibration_readiness},
-    {"demotion.check", handle_demotion_check},
-    {"ranker.export_view", handle_ranker_export_view},
-    {"ranker.fit", handle_ranker_fit},
-    {"workers", handle_workers},
-    /* Rules */
-    {"rules.list", handle_rules_list},
-    {"rules.generate", handle_rules_generate},
-    {"rules.delete", handle_rules_delete},
-    /* Skills */
-    {"skill.list", handle_skill_list},
-    {"skill.show", handle_skill_show},
-    {"skill.lint", handle_skill_lint},
-    {"skill.eval", handle_skill_eval},
-    {"skill.create", handle_skill_create},
-    {"skill.edit", handle_skill_edit},
-    {"skill.patch", handle_skill_patch},
-    {"skill.archive", handle_skill_archive},
-    {"skill.pin", handle_skill_pin},
-    {"skill.unpin", handle_skill_pin},
-    {"skill.lifecycle", handle_skill_lifecycle},
-    {"skill.autostub", handle_skill_autostub},
-    {"collab_rules.list", handle_collab_rules_list},
-    {"collab_rules.list_active", handle_collab_rules_list_active},
-    {"collab_rules.approve", handle_collab_rules_approve},
-    {"collab_rules.reject", handle_collab_rules_reject},
-    {"collab_rules.retire", handle_collab_rules_retire},
-    /* Working memory */
-    {"wm.set", handle_wm_set},
-    {"wm.get", handle_wm_get},
-    {"wm.list", handle_wm_list},
-    {"wm.context", handle_wm_context},
-    /* Per-session primary agent selection */
-    {"primary.set", handle_primary_set},
-    {"hosts.list", handle_hosts_list},
-    {"embedders.list", handle_embedders_list},
-    {"primary.get", handle_primary_get},
-    {"primary.clear", handle_primary_clear},
-    {"attempt.record", handle_attempt_record},
-    {"attempt.list", handle_attempt_list},
-    /* Dashboard */
-    {"dashboard.metrics", handle_dashboard_metrics},
-    {"economizer.stats", handle_economizer_stats},
-    {"dashboard.delegations", handle_dashboard_delegations},
-    {"dashboard.traces", handle_dashboard_traces},
-    {"dashboard.plans", handle_dashboard_plans},
-    {"dashboard.logs", handle_dashboard_logs},
-    {"dashboard.onboard", handle_dashboard_onboard},
-    {"dashboard.memory_stats", handle_dashboard_memory_stats},
-    {"dashboard.all", handle_dashboard_all},
-    {"dashboard.audit", handle_dashboard_audit},
-    {"audit.verify", handle_audit_verify},
-    {"audit.captures", handle_audit_captures},
-    {"audit.replay", handle_audit_replay},
-    {"audit.checkpoint", handle_audit_checkpoint},
-    {"audit.seal", handle_audit_seal},
-    {"audit.snapshot", handle_audit_snapshot},
-    {"lsp.diagnostics_summary", handle_lsp_diagnostics_summary},
-    {"workspace.context", handle_workspace_context},
-    {"workspace.add", handle_workspace_add},
-    {"workspace.list", handle_workspace_list},
-    {"workspace.get", handle_workspace_get},
-    {"workspace.remove", handle_workspace_remove},
-    {"workspace.mirror-sync", handle_workspace_mirror_sync},
-    {"runner.poll", handle_runner_poll},
-    {"runner.respond", handle_runner_respond},
-    {"dogfood.tag", handle_dogfood_tag},
-    {"dogfood.review", handle_dogfood_review},
-    {"dogfood.report", handle_dogfood_report},
-#include "server_dispatch_eval_data.h"
-    /* Identity */
-    {"identity.show", handle_identity_show},
-    {"identity.snapshot", handle_identity_snapshot},
-    {"identity.diff", handle_identity_diff},
-    /* Compute (thread pool) */
-    {"tool.execute", handle_tool_execute},
-    {"delegate", handle_delegate},
-    /* Public /v1 delegate aggregate/roundtable routes enqueue through
-     * rh_dispatch_op_async. Direct raw dispatch remains synchronous for
-     * compatibility with the dispatch-method surface. */
-    {"delegate.aggregate", handle_delegate_aggregate},
-    {"roundtable.review", handle_roundtable_review},
-    {"dev.sweep", handle_dev_sweep},
-    {"delegate.launch", handle_delegate_launch},
-    {"delegate.status", handle_delegate_status},
-    {"delegate.reservation.forget", handle_delegate_reservation_forget},
-    {"delegate.cancel_unassigned", handle_delegate_cancel_unassigned},
-    {"vault.unlock", handle_vault_unlock},
-    {"vault.rekey", handle_vault_rekey},
-    {"vault.set", handle_vault_set},
-    {"vault.set_server", handle_vault_set_server},
-    {"vault.capability", handle_vault_capability},
-    {"vault.list", handle_vault_list},
-    {"vault.delete", handle_vault_delete},
-    {"vault.lock", handle_vault_lock},
-    {"cert.issue", handle_cert_issue},
-    {"cert.sign", handle_cert_sign},
-    {"cert.list", handle_cert_list},
-    {"cert.revoke", handle_cert_revoke},
-    {"jobs.list", handle_jobs_list},
-    {"jobs.status", handle_jobs_status},
-    {"jobs.logs", handle_jobs_logs},
-    {"jobs.cancel", handle_jobs_cancel},
-    {"job.start", handle_coord_job_start},
-    {"job.list", handle_coord_job_list},
-    {"job.status", handle_coord_job_status},
-    {"job.cancel", handle_coord_job_cancel},
-    {"aux.config_show", handle_aux_config_show},
-    {"config.show", handle_config_show},
-    {"config.get", handle_config_get},
-    {"config.deploy_env", handle_config_deploy_env},
-    {"config.set", handle_config_set},
-    {"pipeline.start", handle_pipeline_start},
-    {"pipeline.status", handle_pipeline_status},
-    {"pipeline.list", handle_pipeline_list},
-    {"pipeline.cancel", handle_pipeline_cancel},
-    {"pipeline.resume", handle_pipeline_resume},
-    {"pipeline.advance", handle_pipeline_advance},
-    {"pipeline.gate", handle_pipeline_gate},
-    {"aux.test", handle_aux_test},
-    {"delegate.reply", handle_delegate_reply},
-    {"delegate.log", handle_delegate_log},
-    {"episode.list", handle_episode_list},
-    {"model.episodes", handle_agent_episodes},
-    {"chat.send_stream", handle_chat_send_stream},
-    /* MCP proxy */
-    {"mcp.tools_list", handle_mcp_tools_list},
-    {"mcp.audit", handle_mcp_audit},
-    {"mcp.recheck", handle_mcp_recheck},
-    {"mcp.call", handle_mcp_call},
-    {"help.get", handle_get_help}, /* handler force-selects get_help */
-    /* Triggers */
-    {"trigger.fire", handle_trigger_fire},
-    {"trigger.list", handle_trigger_list},
-    {"trigger.status", handle_trigger_status},
-    {"trigger.cancel", handle_trigger_cancel},
-    {"cron.list", handle_cron_list},
-    {"cron.add", handle_cron_add},
-    {"cron.show", handle_cron_show},
-    {"cron.history", handle_cron_history},
-    {"cron.run", handle_cron_run},
-    {"cron.enable", handle_cron_enable},
-    {"cron.disable", handle_cron_disable},
-    {"cron.remove", handle_cron_remove},
-    {NULL, NULL},
+#include "server/server_dispatch_defs_data.h"
 };
 
 /* --- Dispatch --- */
